+++ src/lib/src/resource-view-registry.spec.ts
@@ -1,12 +1,10 @@
 import { async, inject, TestBed } from '@angular/core/testing';
-import {
-  normalizeStatus,
-  normalizeStatusExpression,
-  RESOURCE_VIEWS,
-  ResourceViewRegistry
-} from './resource-view-registry';
-import { Type } from '@angular/core';
+import { normalizeStatus, normalizeStatusExpression, RESOURCE_VIEWS, ResourceViewRegistry } from './resource-view-registry';
+import { NgModule, Type } from '@angular/core';
 import { ViewDef } from './view-definition';
+import { LoadChildrenCallback } from '@angular/router';
+import { Observable } from 'rxjs';
+import { ResourceRouterModule } from 'angular-resource-router';
 
 describe(ResourceViewRegistry.name, () => {
   beforeEach(async(() => {
@@ -24,7 +22,7 @@ describe(ResourceViewRegistry.name, () => {
   }
 
   function expectComponentName(config: ViewDef | null) {
-    return expect(config ? config.component.toString() : null);
+    return expect(config ? config.component?.toString() : null);
   }
 
   describe('should initialize', () => {
@@ -51,48 +49,97 @@ describe(ResourceViewRegistry.name, () => {
     it('with configured views', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
       expect(registry).toBeTruthy();
       expect(registry.length).toBe(3);
-      expectComponentName(registry.match('one', 200)).toBe('one');
-      expectComponentName(registry.match('two', 200)).toBe('two');
-      expectComponentName(registry.match('three', 200)).toBe('three');
+      expectComponentName(registry.match('one', 200) as ViewDef).toBe('one');
+      expectComponentName(registry.match('two', 200) as ViewDef).toBe('two');
+      expectComponentName(registry.match('three', 200) as ViewDef).toBe('three');
     }));
+  });
+
+  describe('should match lazy loaded component', () => {
+    beforeEach(async(() => {
+      return TestBed.configureTestingModule({
+        providers: [
+          {
+            provide: RESOURCE_VIEWS,
+            useValue: [
+              { type: 'foo', status: 200, component: fakeComponent('200 foo') },
+              { type: 'lazy', status: 200, loadChildren: fakeLoadChildrenCallback() }
+            ],
+            multi: true
+          }
+        ]
+      });
+    }));
+
+    function fakeLoadChildrenCallback(): LoadChildrenCallback {
+      return () => FakeModul;
+    }
+
+    @NgModule({
+      imports: [
+        ResourceRouterModule.forTypes([
+          { type: 'lazy', status: 200, component: fakeComponent('200 lazy') },
+          { type: 'lazy.second', status: 200, component: fakeComponent('200 second') }
+        ])
+      ],
+      exports: [
+        ResourceRouterModule
+      ],
+    })
+    class FakeModul {
+    }
+
+    it('with configured views', done => inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
+      expect(registry).toBeTruthy();
+      expect(registry.length).toBe(2);
+      expectComponentName(registry.match('foo', 200) as ViewDef).toBe('200 foo');
+
+      ( registry.match('lazy', 200) as Observable<ViewDef> )
+        .subscribe(v => {
+          expectComponentName(v).toBe('200 lazy');
+          expectComponentName(registry.match('lazy.second', 200) as ViewDef).toBe('200 second');
+          expect(registry.length).toBe(4);
+          done();
+        });
+    })());
   });
 
   // Exact type
   describe('should match exact type', () => {
     beforeEach(async(inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
       registry.addViews([
-        {type: 'foo', component: fakeComponent('foo')},
-        {type: 'application/foo', component: fakeComponent('application/foo')},
-        {type: 'application/*', component: fakeComponent('application/*')},
-        {type: 'foo', status: 20, component: fakeComponent('020 foo')},
-        {type: 'foo', status: 201, component: fakeComponent('201 foo')},
-        {type: 'foo', status: 400, component: fakeComponent('400 foo')},
-        {type: 'foo', status: '4xx', component: fakeComponent('4?? foo')},
-        {type: 'application/foo', status: '*', component: fakeComponent('??? application/foo')},
+        { type: 'foo', component: fakeComponent('foo') },
+        { type: 'application/foo', component: fakeComponent('application/foo') },
+        { type: 'application/*', component: fakeComponent('application/*') },
+        { type: 'foo', status: 20, component: fakeComponent('020 foo') },
+        { type: 'foo', status: 201, component: fakeComponent('201 foo') },
+        { type: 'foo', status: 400, component: fakeComponent('400 foo') },
+        { type: 'foo', status: '4xx', component: fakeComponent('4?? foo') },
+        { type: 'application/foo', status: '*', component: fakeComponent('??? application/foo') },
       ]);
     })));
 
     it('with default status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foo', 200)).toBe('foo');
-      expectComponentName(registry.match('foo', 299)).toBe('foo');
-      expectComponentName(registry.match('application/foo', 200)).toBe('application/foo');
-      expectComponentName(registry.match('application/foo', 201)).toBe('application/foo');
+      expectComponentName(registry.match('foo', 200) as ViewDef).toBe('foo');
+      expectComponentName(registry.match('foo', 299) as ViewDef).toBe('foo');
+      expectComponentName(registry.match('application/foo', 200) as ViewDef).toBe('application/foo');
+      expectComponentName(registry.match('application/foo', 201) as ViewDef).toBe('application/foo');
     }));
 
     it('with exact status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foo', 201)).toBe('201 foo');
-      expectComponentName(registry.match('foo', 20)).toBe('020 foo');
-      expectComponentName(registry.match('foo', 400)).toBe('400 foo');
+      expectComponentName(registry.match('foo', 201) as ViewDef).toBe('201 foo');
+      expectComponentName(registry.match('foo', 20) as ViewDef).toBe('020 foo');
+      expectComponentName(registry.match('foo', 400) as ViewDef).toBe('400 foo');
     }));
 
     it('with wildcard status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foo', 499)).toBe('4?? foo');
+      expectComponentName(registry.match('foo', 499) as ViewDef).toBe('4?? foo');
     }));
 
     it('with any status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('application/foo', 300)).toBe('??? application/foo');
-      expectComponentName(registry.match('application/foo', 0)).toBe('??? application/foo');
-      expectComponentName(registry.match('application/foo', 999)).toBe('??? application/foo');
+      expectComponentName(registry.match('application/foo', 300) as ViewDef).toBe('??? application/foo');
+      expectComponentName(registry.match('application/foo', 0) as ViewDef).toBe('??? application/foo');
+      expectComponentName(registry.match('application/foo', 999) as ViewDef).toBe('??? application/foo');
     }));
 
     it('without definition as null', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
@@ -112,9 +159,9 @@ describe(ResourceViewRegistry.name, () => {
       });
 
       // Test
-      expectComponentName(registry.match('foo', 200)).toBe('quality foo');
-      expectComponentName(registry.match('foo', 201)).toBe('201 quality foo');
-      expectComponentName(registry.match('application/foo', 200)).toBe('application/foo');
+      expectComponentName(registry.match('foo', 200) as ViewDef).toBe('quality foo');
+      expectComponentName(registry.match('foo', 201) as ViewDef).toBe('201 quality foo');
+      expectComponentName(registry.match('application/foo', 200) as ViewDef).toBe('application/foo');
     }));
   });
 
@@ -139,44 +186,44 @@ describe(ResourceViewRegistry.name, () => {
     })));
 
     it('with default status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foo', 200)).toBe('foo*');
-      expectComponentName(registry.match('foobar', 200)).toBe('foo*');
-      expectComponentName(registry.match('foobar', 299)).toBe('foo*');
-      expectComponentName(registry.match('fubar', 200)).toBe('*bar');
-      expectComponentName(registry.match('application/foo', 200)).toBe('application/f?o');
-      expectComponentName(registry.match('application/fxo', 200)).toBe('application/f?o');
-      expectComponentName(registry.match('application/fubar', 202)).toBe('*bar');
-      expectComponentName(registry.match('application/fooxxxbar', 200)).toBe('*bar');
-      expectComponentName(registry.match('application/bar', 200)).toBe('*bar');
-      expectComponentName(registry.match('application/xxx', 299)).toBe('application/*');
+      expectComponentName(registry.match('foo', 200) as ViewDef).toBe('foo*');
+      expectComponentName(registry.match('foobar', 200) as ViewDef).toBe('foo*');
+      expectComponentName(registry.match('foobar', 299) as ViewDef).toBe('foo*');
+      expectComponentName(registry.match('fubar', 200) as ViewDef).toBe('*bar');
+      expectComponentName(registry.match('application/foo', 200) as ViewDef).toBe('application/f?o');
+      expectComponentName(registry.match('application/fxo', 200) as ViewDef).toBe('application/f?o');
+      expectComponentName(registry.match('application/fubar', 202) as ViewDef).toBe('*bar');
+      expectComponentName(registry.match('application/fooxxxbar', 200) as ViewDef).toBe('*bar');
+      expectComponentName(registry.match('application/bar', 200) as ViewDef).toBe('*bar');
+      expectComponentName(registry.match('application/xxx', 299) as ViewDef).toBe('application/*');
     }));
 
     it('with exact status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foobar', 201)).toBe('201 *bar');
-      expectComponentName(registry.match('application/fubar', 201)).toBe('201 *bar');
-      expectComponentName(registry.match('foobar', 20)).toBe('020 foo*');
-      expectComponentName(registry.match('foo', 400)).toBe('400 foo*');
-      expectComponentName(registry.match('fooxxx', 400)).toBe('400 foo*');
-      expectComponentName(registry.match('brr', 200)).toBe('200 b?r');
+      expectComponentName(registry.match('foobar', 201) as ViewDef).toBe('201 *bar');
+      expectComponentName(registry.match('application/fubar', 201) as ViewDef).toBe('201 *bar');
+      expectComponentName(registry.match('foobar', 20) as ViewDef).toBe('020 foo*');
+      expectComponentName(registry.match('foo', 400) as ViewDef).toBe('400 foo*');
+      expectComponentName(registry.match('fooxxx', 400) as ViewDef).toBe('400 foo*');
+      expectComponentName(registry.match('brr', 200) as ViewDef).toBe('200 b?r');
     }));
 
     it('with wildcard status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foo', 499)).toBe('4?? foo*');
-      expectComponentName(registry.match('foobar', 410)).toBe('4?? foo*');
-      expectComponentName(registry.match('fooxxx', 401)).toBe('4?? foo*');
+      expectComponentName(registry.match('foo', 499) as ViewDef).toBe('4?? foo*');
+      expectComponentName(registry.match('foobar', 410) as ViewDef).toBe('4?? foo*');
+      expectComponentName(registry.match('fooxxx', 401) as ViewDef).toBe('4?? foo*');
     }));
 
     it('with any status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('application/foo', 300)).toBe('??? application/foo*');
-      expectComponentName(registry.match('application/foobar', 300)).toBe('??? application/foo*');
-      expectComponentName(registry.match('application/fooxxx', 300)).toBe('??? application/foo*');
-      expectComponentName(registry.match('application/foo', 0)).toBe('??? application/foo*');
-      expectComponentName(registry.match('application/foobar', 0)).toBe('??? application/foo*');
-      expectComponentName(registry.match('application/fooxxx', 0)).toBe('??? application/foo*');
-      expectComponentName(registry.match('application/fubar', 999)).toBe('??? application/*bar');
-      expectComponentName(registry.match('application/fubar', 400)).toBe('??? application/*bar');
-      expectComponentName(registry.match('image/png', 200)).toBe('??? image/*');
-      expectComponentName(registry.match('image/jpg', 400)).toBe('??? image/*');
+      expectComponentName(registry.match('application/foo', 300) as ViewDef).toBe('??? application/foo*');
+      expectComponentName(registry.match('application/foobar', 300) as ViewDef).toBe('??? application/foo*');
+      expectComponentName(registry.match('application/fooxxx', 300) as ViewDef).toBe('??? application/foo*');
+      expectComponentName(registry.match('application/foo', 0) as ViewDef).toBe('??? application/foo*');
+      expectComponentName(registry.match('application/foobar', 0) as ViewDef).toBe('??? application/foo*');
+      expectComponentName(registry.match('application/fooxxx', 0) as ViewDef).toBe('??? application/foo*');
+      expectComponentName(registry.match('application/fubar', 999) as ViewDef).toBe('??? application/*bar');
+      expectComponentName(registry.match('application/fubar', 400) as ViewDef).toBe('??? application/*bar');
+      expectComponentName(registry.match('image/png', 200) as ViewDef).toBe('??? image/*');
+      expectComponentName(registry.match('image/jpg', 400) as ViewDef).toBe('??? image/*');
     }));
 
     it('without definition as Error', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
@@ -191,18 +238,18 @@ describe(ResourceViewRegistry.name, () => {
     it('with quality override', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
       // Extra data
       registry.addViews([
-        {type: 'foo*', quality: 0.51, component: fakeComponent('quality foo*')},
-        {type: '*bar', quality: 0.6, component: fakeComponent('quality *bar')},
-        {type: 'foo*', quality: 0.51, status: 201, component: fakeComponent('201 quality foo*')},
-        {type: 'application/foo*', quality: 0.49, component: fakeComponent('low quality foo*')}
+        { type: 'foo*', quality: 0.51, component: fakeComponent('quality foo*') },
+        { type: '*bar', quality: 0.6, component: fakeComponent('quality *bar') },
+        { type: 'foo*', quality: 0.51, status: 201, component: fakeComponent('201 quality foo*') },
+        { type: 'application/foo*', quality: 0.49, component: fakeComponent('low quality foo*') }
       ]);
 
       // Test
-      expectComponentName(registry.match('foo', 200)).toBe('quality foo*');
-      expectComponentName(registry.match('foobar', 200)).toBe('quality *bar');
-      expectComponentName(registry.match('foo', 201)).toBe('201 quality foo*');
-      expectComponentName(registry.match('foobar', 201)).toBe('201 quality foo*');
-      expectComponentName(registry.match('application/foo', 200)).toBe('application/f?o');
+      expectComponentName(registry.match('foo', 200) as ViewDef).toBe('quality foo*');
+      expectComponentName(registry.match('foobar', 200) as ViewDef).toBe('quality *bar');
+      expectComponentName(registry.match('foo', 201) as ViewDef).toBe('201 quality foo*');
+      expectComponentName(registry.match('foobar', 201) as ViewDef).toBe('201 quality foo*');
+      expectComponentName(registry.match('application/foo', 200) as ViewDef).toBe('application/f?o');
     }));
   });
 
@@ -220,44 +267,44 @@ describe(ResourceViewRegistry.name, () => {
     })));
 
     it('with default status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foo', 200)).toBe('*');
-      expectComponentName(registry.match('asd', 299)).toBe('*');
-      expectComponentName(registry.match('application/foo', 200)).toBe('*');
-      expectComponentName(registry.match('', 200)).toBe('*');
+      expectComponentName(registry.match('foo', 200) as ViewDef).toBe('*');
+      expectComponentName(registry.match('asd', 299) as ViewDef).toBe('*');
+      expectComponentName(registry.match('application/foo', 200) as ViewDef).toBe('*');
+      expectComponentName(registry.match('', 200) as ViewDef).toBe('*');
     }));
 
     it('with exact status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foobar', 201)).toBe('201 *');
-      expectComponentName(registry.match('application/fubar', 201)).toBe('201 *');
-      expectComponentName(registry.match('asd', 20)).toBe('020 *');
-      expectComponentName(registry.match('aaa', 400)).toBe('400 *');
+      expectComponentName(registry.match('foobar', 201) as ViewDef).toBe('201 *');
+      expectComponentName(registry.match('application/fubar', 201) as ViewDef).toBe('201 *');
+      expectComponentName(registry.match('asd', 20) as ViewDef).toBe('020 *');
+      expectComponentName(registry.match('aaa', 400) as ViewDef).toBe('400 *');
     }));
 
     it('with wildcard status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('foo', 499)).toBe('4?? *');
-      expectComponentName(registry.match('foobar', 410)).toBe('4?? *');
-      expectComponentName(registry.match('fooxxx', 401)).toBe('4?? *');
+      expectComponentName(registry.match('foo', 499) as ViewDef).toBe('4?? *');
+      expectComponentName(registry.match('foobar', 410) as ViewDef).toBe('4?? *');
+      expectComponentName(registry.match('fooxxx', 401) as ViewDef).toBe('4?? *');
     }));
 
     it('with any status', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
-      expectComponentName(registry.match('application/foo', 300)).toBe('??? *');
-      expectComponentName(registry.match('foo', 999)).toBe('??? *');
-      expectComponentName(registry.match('application', 0)).toBe('??? *');
+      expectComponentName(registry.match('application/foo', 300) as ViewDef).toBe('??? *');
+      expectComponentName(registry.match('foo', 999) as ViewDef).toBe('??? *');
+      expectComponentName(registry.match('application', 0) as ViewDef).toBe('??? *');
     }));
 
     it('with quality override', inject([ResourceViewRegistry], (registry: ResourceViewRegistry) => {
       // Extra data
       registry.addViews([
-        {type: '*', quality: 0.01, component: fakeComponent('quality *')},
-        {type: '*', quality: 0.01, status: 201, component: fakeComponent('201 quality *')},
-        {type: '*', quality: -1, component: fakeComponent('low quality *')}
+        { type: '*', quality: 0.01, component: fakeComponent('quality *') },
+        { type: '*', quality: 0.01, status: 201, component: fakeComponent('201 quality *') },
+        { type: '*', quality: -1, component: fakeComponent('low quality *') }
       ]);
 
       // Test
-      expectComponentName(registry.match('foo', 200)).toBe('quality *');
-      expectComponentName(registry.match('xxx', 201)).toBe('201 quality *');
-      expectComponentName(registry.match('foo', 400)).toBe('400 *');
-      expectComponentName(registry.match('foo', 410)).toBe('4?? *');
+      expectComponentName(registry.match('foo', 200) as ViewDef).toBe('quality *');
+      expectComponentName(registry.match('xxx', 201) as ViewDef).toBe('201 quality *');
+      expectComponentName(registry.match('foo', 400) as ViewDef).toBe('400 *');
+      expectComponentName(registry.match('foo', 410) as ViewDef).toBe('4?? *');
     }));
   });
 
@@ -270,9 +317,9 @@ describe(ResourceViewRegistry.name, () => {
 
     // Test
     expect(registry.length).toBe(9);
-    expectComponentName(registry.match('foo', 200)).toBeTruthy();
-    expectComponentName(registry.match('bar', 200)).toBeTruthy();
-    expectComponentName(registry.match('foobar', 501)).toBeTruthy();
+    expectComponentName(registry.match('foo', 200) as ViewDef).toBeTruthy();
+    expectComponentName(registry.match('bar', 200) as ViewDef).toBeTruthy();
+    expectComponentName(registry.match('foobar', 501) as ViewDef).toBeTruthy();
     expect(() => registry.match('xxx', 200)).toThrow();
   }));